requirements.txt
torch==1.5.0
torchvision==0.6.0
pretrainedmodels==0.7.4
efficientnet-pytorch==0.6.3
pandas==1.0.4
statsmodels==0.11.1
pillow==7.1.2
opencv-python==4.2.0.34
seaborn==0.10.1
protobuf==3.12.2
psycopg2-binary==2.8.5
Flask-SQLAlchemy==2.4.3
SQLAlchemy-Utils==0.36.6
nvidia-ml-py3==7.352.0
mplcursors==0.3
boto3==1.13.24
botocore==1.16.24
ipython==7.15.0
click==7.1.1
timm==0.1.28
plotly==4.8.2
git+https://github.com/tensorflow/cleverhans.git#egg=cleverhans
